--- setup.py
@@ -1,3 +1,0 @@
-from setuptools import setup, find_packages
-
-setup(name='hific', version='0.0.1', packages=find_packages())
