+++ src/normalisation/instance.py
@@ -1,0 +1,15 @@
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+
+from torch.nn import Parameter
+
+def InstanceNorm2D_wrap(input_channels, momentum=0.1, affine=True,
+                        track_running_stats=False, **kwargs):
+    """ 
+    Wrapper around default Torch instancenorm
+    """
+    instance_norm_layer = nn.InstanceNorm2d(input_channels, 
+        momentum=momentum, affine=affine,
+        track_running_stats=track_running_stats)
+    return instance_norm_layer
