+++ src/perceptual_similarity/dist_model.py
@@ -17,7 +17,7 @@ import skimage.transform
 from tqdm import tqdm
 
 from . import networks_basic as networks
-from hific import perceptual_similarity as util
+import util
 
 class DistModel(BaseModel):
     def name(self):
